pyproject.toml
[project]
name = "auth-challenge"
version = "0.1.0"
description = "Minimal DDD/CQRS authentication challenge implementation"
requires-python = ">=3.14"
dependencies = [
  "fastapi==0.135.1",
  "strawberry-graphql[fastapi]==0.307.1",
  "uvicorn==0.41.0",
  "sqlalchemy==2.0.48",
  "asyncpg==0.31.0",
  "PyJWT==2.11.0",
  "passlib==1.7.4",
]

[dependency-groups]
dev = [
  "pytest==9.0.2",
  "pytest-asyncio==1.3.0",
  "pytest-cov==7.0.0",
  "aiosqlite==0.22.1",
  "httpx==0.28.1",
  "ruff==0.14.14",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py314"
src = ["src", "tests"]
exclude = [
  "app",
  "cfg",
  "pydantic_quest",
  ".venv",
]

[tool.ruff.lint]
select = ["FURB", "W", "PLW", "RUF", "E", "F", "B", "N", "UP", "SIM", "C4", "T20", "ARG", "ASYNC", "ANN", "FBT", "COM", "C4", "PLC", "DTZ", "EM", "PIE", "T20", "PT", "RSE", "RET", "TRY"]
ignore = [
  "E501",
  "B008",
  "RET504",
  "W293",
  "COM812",
]

[tool.ruff.format]

[tool.coverage.run]
source = ["src"]
